+++ CHANGELOG.md
@@ -98,7 +98,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * `torchmetrics.FBeta` -> `torchmetrics.FBetaScore`
 
 
-- Renamed pit metric: ([#737](https://github.com/PyTorchLightning/metrics/pull/737))
+- Renamed image metrics ([#732](https://github.com/PyTorchLightning/metrics/pull/732))
+  * `functional.psnr` -> `functional.peak_signal_noise_ratio`
+  * `PSNR` -> `PeakSignalNoiseRatio`
+
+
+- Renamed PIT metric: ([#737](https://github.com/PyTorchLightning/metrics/pull/737))
   * `torchmetrics.functional.pit` -> `torchmetrics.functional.permutation_invariant_training`
   * `torchmetrics.PIT` -> `torchmetrics.PermutationInvariantTraining`